// moviesapp/src/app/constacts/sweetalert-enum.ts
export enum MessageType {
    Error = 'error',
    Message = 'messageTitle',
    Notify = 'notify',
    Success = 'success',
    Warning = 'warning',
  }

export enum Position {
    TopCenter = "top-center",
    TopRight = "top-right",
    TopLeft = "top-left",
    Center = "center",
    BottomRight = "bottom-right",
    BottomCenter = "bottom-center",
    BottomLeft = "bottom-left"
}